Added som more...
Added a bodeplot, a EBP tool, a bunch of equation snippets to the
transfer function of a vented box preassure respone, both butterworth
and chebyshev. Not sure how th etransfer function should be translated
and put into the scipy.signal.lti(). Plotted the system but not
convinced of the outcome. Had a look at the scimpy module to try and get
some more insight.